comments - no longer needed? Plus add int() call QGL version added

+++ src/python/qgl2/basic_sequences/pulses.py
@@ -6,12 +6,13 @@ import numpy as np
 
 # a local copy of QGL.PulseShapes.tanh, because pulling in
 # QGL/__init__.py causes QGL2 grief.
-# See RabiMin
+# See RabiWidth
+# - FIXME: No longer needed?
 def local_tanh(amp=1, length=0, sigma=0, cutoff=2, sampling_rate=1e9, **params):
     '''
     A rounded constant shape from the sum of two tanh shapes.
     '''
-    numPts = np.round(length * sampling_rate)
+    numPts = int(np.round(length * sampling_rate))
     xPts = np.linspace(-length / 2, length / 2, numPts)
     x1 = -length / 2 + cutoff * sigma
     x2 = +length / 2 - cutoff * sigma
